Fix/tag question (#28)
* Refactors Question-Tag relationship to one-to-many

Changes the relationship between Question and Tag entities from many-to-many to one-to-many.

Updates the database configuration to reflect the change, including setting up a foreign key and cascade delete.

Removes unnecessary many-to-many configuration from the Tag entity.

* Refactors question domain and persistence

Removes the `Tag` entity and replaces it with `Level`, `Content`, and `SubContent` entities for better question categorization.

Updates the `Question` entity to include references to the new categorization entities.

Adds configurations and migrations for the new entities.